JG-2003-05-21-0  Bugfix:    SGE didn't compile on systems with glibc-2.3 and                             didn't compile with gcc-3.3                  Review:    EB                  Changed:   all Issue number: Obtained from: Submitted by: Reviewed by:

+++ source/libs/gdi/gdi_checkpermissions.c
@@ -113,7 +113,7 @@ int buflen
    }
   
    if (mapName != NULL) {
-      if ((strlen(mapName) + 1) <= buflen) {
+      if ((strlen(mapName) + 1) <= (size_t)buflen) {
          strcpy(buf,mapName);
          DPRINTF(("Mapping name is: '%s'\n", buf));
